Add structure for HTML for first column of page
My basic plan here is to tackle both project sections independently. So
I want to get the first column with the background image fully finished
before creating the form column on the right side of the page. Reminder
to put a link to the project page so that the design reference can be
viewed.