Finish Left Column, Add framework for Right Column
Reminder to look up how to deal with collapsing margins. Also concerned
about how I am going to format the form correctly with the way it is
set up in the reference image but for now I just want to get all the
elements in place and then worry about how I am going to fix it from
there.