Add form structure, add all elements
I could use fewer classes here by putting the button on the outside of
the form but I think in a real world context I would want that button
associated with the form so I'm going to keep it this way for now. The
trick now is to figure out how to style everything properly.